requirements.txt
requests==2.31.0
python-dotenv==1.0.0
pandas==2.1.1
streamlit==1.27.2
plotly==5.17.0